aligned classes structure
migrate asn1 class functions to dynamic functions
