+++ src/PDFDoc.php
@@ -31,8 +31,8 @@ use ddn\sapp\pdfvalue\PDFValueSimple;
 use ddn\sapp\pdfvalue\PDFValueHexString;
 use ddn\sapp\pdfvalue\PDFValueString;
 use ddn\sapp\helpers\CMS;
-// use ddn\sapp\helpers\x509;
-// use ddn\sapp\helpers\asn1;
+use ddn\sapp\helpers\x509;
+use ddn\sapp\helpers\asn1;
 use ddn\sapp\helpers\Buffer;
 use ddn\sapp\helpers\UUID;
 use ddn\sapp\helpers\DependencyTreeObject;
